requirements.txt
streamlit==1.30.0
python-dotenv==1.1.1
langchain-ollama==0.2.0
langchain-core==0.3.74
langchain-community==0.3.27
ollama==0.4.6
langchain-huggingface==0.1.0
langchain-openai==0.2.3
langsmith==0.4.14
langgraph==0.2.43
faiss-cpu==1.9.0
sentence-transformers==5.1.0
